fix recursion with conditional check before updating filterScores
